pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rembg-pipeline"
version = "2.0.0"
description = "Production-ready multi-stage background removal pipeline"
requires-python = ">=3.8"
dependencies = [
    "opencv-python",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "scipy>=1.11.0",
]

[project.scripts]
rembg = "rembg.cli:main"
